server/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-multipart
jinja2
supabase
python-dotenv
asyncpg
psycopg2-binary
